[lc_ctrl,dv] Document last functions in lc_ctrl.h
Also, tighten the return type of read_lc_counter (since it never
returns anything bigger than 0x1f).

+++ src/integration/test_suites/libs/lc_ctrl/lc_ctrl.c
@@ -393,7 +393,7 @@ bool check_lc_state(const char *desc, uint8_t exp_idx)
     return true;
 }
 
-uint32_t read_lc_counter(void) {
+uint8_t read_lc_counter(void) {
     mcu_sleep(512);
 
     // Read LC_CTRL_LC_TRANSITION_CNT register and mask out the reserved bits (bits 31:5)
